+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changelog
 0.10 (unreleased)
 -----------------
 
+- Shop Portlet: Use appropriate context for "orders in context", like it's done
+  for "bookings in context". Avoids calling this view on default pages.
+  [thet]
+
 - Shop Portlet: Add Shop Controlpanel admin link, linking to ISite or
   IPloneSiteRoot. If lineage.registry is installed, the shop settings can also
   be set on subsite level.